requirements.txt
ipaddr==2.2.0
netifaces==0.10.7
python-nmap==0.6.1
pexpect==4.3.1
